revert(content): inject content programmically
