+++ src/manifest.json
@@ -15,8 +15,14 @@
     "default_popup": "popup/index.html",
     "default_title": "Comment Wizard"
   },
-  "permissions": ["storage", "tabs", "activeTab", "scripting"],
+  "permissions": ["storage"],
   "host_permissions": ["https://www.webtoons.com/"],
+  "content_scripts": [
+    {
+      "matches": ["https://www.webtoons.com/*"],
+      "js": ["content.js"]
+    }
+  ],
   "background": {
     "service_worker": "worker.js",
     "type": "module"
